Apply dhewm3 / 0d27bfe7b9337cb1a311ce4a67c43736f9495c6c: Remove usage of long type from idlib/
Apart from some minor stuff, this changes the signature of some methods of Parser and Token classes and of the (unused) Random2 class.
That no problem though, because the calling code uses normal ints anyway.


git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/branches/x64@6285 49c82d7f-2e2a-0410-a16f-ae8f201b507f

+++ idlib/Parser.cpp
@@ -1285,7 +1285,7 @@ typedef struct operator_s
 
 typedef struct value_s
 {
-	signed long int intvalue;
+	int intvalue;
 	double floatvalue;
 	int parentheses;
 	struct value_s *prev, *next;
@@ -1356,8 +1356,8 @@ int PC_OperatorPriority(int op) {
 
 #define FreeOperator(op)
 
-int idParser::EvaluateTokens( idToken *tokens, signed long int *intvalue, double *floatvalue, int integer ) {
-	operator_t *o, *firstoperator, *lastoperator;
+int idParser::EvaluateTokens( idToken *tokens, int *intvalue, double *floatvalue, int integer ) {
+    operator_t *o, *firstoperator, *lastoperator;
 	value_t *v, *firstvalue, *lastvalue, *v1, *v2;
 	idToken *t;
 	int brace = 0;
@@ -1781,7 +1781,7 @@ int idParser::EvaluateTokens( idToken *tokens, signed long int *intvalue, double
 idParser::Evaluate
 ================
 */
-int idParser::Evaluate( signed long int *intvalue, double *floatvalue, int integer ) {
+int idParser::Evaluate( int *intvalue, double *floatvalue, int integer ) {
 	idToken token, *firsttoken, *lasttoken;
 	idToken *t, *nexttoken;
 	define_t *define;
@@ -1872,8 +1872,8 @@ int idParser::Evaluate( signed long int *intvalue, double *floatvalue, int integ
 idParser::DollarEvaluate
 ================
 */
-int idParser::DollarEvaluate( signed long int *intvalue, double *floatvalue, int integer) {
-	int indent, defined = false;
+int idParser::DollarEvaluate( int *intvalue, double *floatvalue, int integer) {
+    int indent, defined = false;
 	idToken token, *firsttoken, *lasttoken;
 	idToken *t, *nexttoken;
 	define_t *define;
@@ -1974,7 +1974,7 @@ idParser::Directive_elif
 ================
 */
 int idParser::Directive_elif( void ) {
-	signed long int value;
+	int value;
 	int type, skip;
 
 	idParser::PopIndent( &type, &skip );
@@ -1996,7 +1996,7 @@ idParser::Directive_if
 ================
 */
 int idParser::Directive_if( void ) {
-	signed long int value;
+	int value;
 	int skip;
 
 	if ( !idParser::Evaluate( &value, NULL, true ) ) {
@@ -2092,7 +2092,7 @@ idParser::Directive_eval
 ================
 */
 int idParser::Directive_eval( void ) {
-	signed long int value;
+	int value;
 	idToken token;
 	char buf[128];
 
@@ -2231,7 +2231,7 @@ idParser::DollarDirective_evalint
 ================
 */
 int idParser::DollarDirective_evalint( void ) {
-	signed long int value;
+	int value;
 	idToken token;
 	char buf[128];
 
@@ -2280,7 +2280,7 @@ int idParser::DollarDirective_evalfloat( void ) {
 	token = buf;
 	token.type = TT_NUMBER;
 	token.subtype = TT_FLOAT | TT_LONG | TT_DECIMAL | TT_VALUESVALID;
-	token.intvalue = (unsigned long) fabs( value );
+	token.intvalue = (unsigned int) fabs( value );
 	token.floatvalue = fabs( value );
 	idParser::UnreadSourceToken( &token );
 	if ( value < 0 ) {
